api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
strawberry-graphql[fastapi]==0.227.0
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
redis[asyncio]==5.0.4
aiokafka==0.10.0
